+++ src/main/java/tenno_mod/cards/common/CombatKnowledge_TENNO.java
@@ -1,0 +1,86 @@
+package tenno_mod.cards.common;
+
+import basemod.abstracts.CustomCard;
+import com.megacrit.cardcrawl.actions.AbstractGameAction;
+import com.megacrit.cardcrawl.actions.common.DamageAction;
+import com.megacrit.cardcrawl.actions.common.DrawCardAction;
+import com.megacrit.cardcrawl.actions.common.GainBlockAction;
+import com.megacrit.cardcrawl.cards.AbstractCard;
+import com.megacrit.cardcrawl.cards.DamageInfo;
+import com.megacrit.cardcrawl.characters.AbstractPlayer;
+import com.megacrit.cardcrawl.core.CardCrawlGame;
+import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
+import com.megacrit.cardcrawl.localization.CardStrings;
+import com.megacrit.cardcrawl.monsters.AbstractMonster;
+import tenno_mod.patches.AbstractCardEnum;
+
+public class CombatKnowledge_TENNO extends CustomCard {
+  public static final String ID = "CombatKnowledge_TENNO";
+  private static final CardStrings cardStrings = CardCrawlGame.languagePack.getCardStrings(ID);
+  public static final String NAME = cardStrings.NAME;
+  public static final String DESCRIPTION = cardStrings.DESCRIPTION;
+  public static final String UPGRADE_DESCRIPTION = cardStrings.UPGRADE_DESCRIPTION;
+  public static final String[] EXTENDED_DESCRIPTION = cardStrings.EXTENDED_DESCRIPTION;
+  public static final String IMG_PATH = "img/cards/Skill.png";
+  private static final int COST = 1;
+  private static final int MAGIC_NUMBER = 2;
+
+  public CombatKnowledge_TENNO() {
+    super(
+        ID,
+        NAME,
+        IMG_PATH,
+        COST,
+        DESCRIPTION,
+        CardType.SKILL,
+        AbstractCardEnum.TENNO_COLOR,
+        CardRarity.COMMON,
+        CardTarget.SELF
+    );
+    this.baseBlock = 0;
+    this.magicNumber = this.baseMagicNumber = MAGIC_NUMBER;
+  }
+
+  public void use(AbstractPlayer p, AbstractMonster m) {
+    AbstractDungeon.actionManager.addToBottom(
+        new DrawCardAction(p, this.magicNumber));
+    AbstractDungeon.actionManager.addToBottom(
+        new GainBlockAction(p, p, this.block)
+    );
+  }
+
+  public AbstractCard makeCopy() {
+    return new CombatKnowledge_TENNO();
+  }
+
+  public void applyPowers() {
+    AbstractPlayer p = AbstractDungeon.player;
+    this.baseBlock = p.hand.size();
+    int drawableCards = p.drawPile.size() + p.discardPile.size();
+    if (drawableCards > 2) {
+      this.baseBlock += 2;
+    } else {
+      this.baseBlock += drawableCards;
+    }
+    if (this.upgraded) {
+      this.baseBlock +=3;
+    }
+    super.applyPowers();
+    if (!this.upgraded) {
+      this.rawDescription = DESCRIPTION;
+    } else {
+      this.rawDescription = UPGRADE_DESCRIPTION;
+    }
+    this.rawDescription += EXTENDED_DESCRIPTION[0];
+    initializeDescription();
+  }
+
+  public void upgrade() {
+    if (!this.upgraded) {
+      upgradeName();
+      upgradeBlock(3);
+      this.rawDescription = UPGRADE_DESCRIPTION;
+      initializeDescription();
+    }
+  }
+}