+++ src/main/java/tenno_mod/cards/rare/MasterfulCombination_TENNO.java
@@ -1,0 +1,55 @@
+package tenno_mod.cards.rare;
+
+import basemod.abstracts.CustomCard;
+import com.megacrit.cardcrawl.actions.unique.AttackFromDeckToHandAction;
+import com.megacrit.cardcrawl.actions.unique.SkillFromDeckToHandAction;
+import com.megacrit.cardcrawl.cards.AbstractCard;
+import com.megacrit.cardcrawl.characters.AbstractPlayer;
+import com.megacrit.cardcrawl.core.CardCrawlGame;
+import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
+import com.megacrit.cardcrawl.localization.CardStrings;
+import com.megacrit.cardcrawl.monsters.AbstractMonster;
+import tenno_mod.patches.AbstractCardEnum;
+
+public class MasterfulCombination_TENNO extends CustomCard {
+  public static final String ID = "MasterfulCombination_TENNO";
+  private static final CardStrings cardStrings = CardCrawlGame.languagePack.getCardStrings(ID);
+  public static final String NAME = cardStrings.NAME;
+  public static final String DESCRIPTION = cardStrings.DESCRIPTION;
+  public static final String IMG_PATH = "img/cards/Skill.png";
+  private static final int COST = 1;
+  private static final int UPG_COST = 0;
+
+  public MasterfulCombination_TENNO() {
+    super(
+        ID,
+        NAME,
+        IMG_PATH,
+        COST,
+        DESCRIPTION,
+        CardType.SKILL,
+        AbstractCardEnum.TENNO_COLOR,
+        CardRarity.RARE,
+        CardTarget.SELF
+    );
+    this.exhaust = true;
+  }
+
+  public void use(AbstractPlayer p, AbstractMonster m) {
+    AbstractDungeon.actionManager.addToBottom(
+        new AttackFromDeckToHandAction(1));
+    AbstractDungeon.actionManager.addToBottom(
+        new SkillFromDeckToHandAction(1));
+  }
+
+  public AbstractCard makeCopy() {
+    return new MasterfulCombination_TENNO();
+  }
+
+  public void upgrade() {
+    if (!this.upgraded) {
+      upgradeName();
+      upgradeBaseCost(UPG_COST);
+    }
+  }
+}